refactor: split tui render surfaces and add runtime hints
