+++ internal/detect/ollama.go
@@ -45,16 +45,21 @@ func detectOllama(result *Result) {
 		ExtraData: make(map[string]string),
 	}
 
+	acct.SetHint("config_dir", configDir)
+	acct.SetHint("cloud_base_url", "https://ollama.com")
 	acct.ExtraData["config_dir"] = configDir
 	acct.ExtraData["cloud_base_url"] = "https://ollama.com"
 
 	if fileExists(dbPath) {
+		acct.SetHint("db_path", dbPath)
 		acct.ExtraData["db_path"] = dbPath
 	}
 	if dirExists(logsDir) {
+		acct.SetHint("logs_dir", logsDir)
 		acct.ExtraData["logs_dir"] = logsDir
 	}
 	if fileExists(serverConfig) {
+		acct.SetHint("server_config", serverConfig)
 		acct.ExtraData["server_config"] = serverConfig
 	}
