+++ internal/tui/detail.go
@@ -2,8 +2,6 @@ package tui
 
 import (
 	"fmt"
-	"math"
-	"strconv"
 	"strings"
 	"time"
 
@@ -308,356 +306,4 @@ func renderDetailHeader(sb *strings.Builder, snap core.UsageSnapshot, w int) {
 
 	sb.WriteString(card)
 	sb.WriteString("\n")
-}
-
-func wrapTags(tags []string, maxWidth int) []string {
-	if len(tags) == 0 {
-		return nil
-	}
-	var rows []string
-	currentRow := ""
-	currentW := 0
-	sep := " "
-	sepW := 1
-
-	for _, tag := range tags {
-		tagW := lipgloss.Width(tag)
-		if currentW > 0 && currentW+sepW+tagW > maxWidth {
-			rows = append(rows, currentRow)
-			currentRow = tag
-			currentW = tagW
-		} else {
-			if currentW > 0 {
-				currentRow += sep
-				currentW += sepW
-			}
-			currentRow += tag
-			currentW += tagW
-		}
-	}
-	if currentRow != "" {
-		rows = append(rows, currentRow)
-	}
-	return rows
-}
-
-func titleCase(s string) string {
-	if len(s) <= 1 {
-		return s
-	}
-	return strings.ToUpper(s[:1]) + strings.ToLower(s[1:])
-}
-
-// renderInfoSection renders Attributes, Diagnostics, and Raw as separate sub-sections.
-func renderInfoSection(sb *strings.Builder, snap core.UsageSnapshot, widget core.DashboardWidget, w int) {
-	labelW := sectionLabelWidth(w)
-	maxValW := w - labelW - 6
-	if maxValW < 20 {
-		maxValW = 20
-	}
-	if maxValW > 45 {
-		maxValW = 45
-	}
-
-	if len(snap.Attributes) > 0 {
-		renderDetailSectionHeader(sb, "Attributes", w)
-		renderKeyValuePairs(sb, snap.Attributes, labelW, maxValW, valueStyle)
-	}
-
-	if len(snap.Diagnostics) > 0 {
-		if len(snap.Attributes) > 0 {
-			sb.WriteString("\n")
-		}
-		renderDetailSectionHeader(sb, "Diagnostics", w)
-		warnValueStyle := lipgloss.NewStyle().Foreground(colorYellow)
-		renderKeyValuePairs(sb, snap.Diagnostics, labelW, maxValW, warnValueStyle)
-	}
-
-	if len(snap.Raw) > 0 {
-		if len(snap.Attributes) > 0 || len(snap.Diagnostics) > 0 {
-			sb.WriteString("\n")
-		}
-		renderDetailSectionHeader(sb, "Raw Data", w)
-		renderRawData(sb, snap.Raw, widget, w)
-	}
-}
-
-// renderKeyValuePairs renders a sorted key-value map with consistent formatting.
-func renderKeyValuePairs(sb *strings.Builder, data map[string]string, labelW, maxValW int, vs lipgloss.Style) {
-	keys := core.SortedStringKeys(data)
-
-	for _, k := range keys {
-		v := smartFormatValue(data[k])
-		if len(v) > maxValW {
-			v = v[:maxValW-3] + "..."
-		}
-		sb.WriteString(fmt.Sprintf("  %s  %s\n",
-			labelStyle.Width(labelW).Render(prettifyKey(k)),
-			vs.Render(v),
-		))
-	}
-}
-
-func renderRawData(sb *strings.Builder, raw map[string]string, widget core.DashboardWidget, w int) {
-	labelW := sectionLabelWidth(w)
-
-	maxValW := w - labelW - 6
-	if maxValW < 20 {
-		maxValW = 20
-	}
-	if maxValW > 45 {
-		maxValW = 45
-	}
-
-	rendered := make(map[string]bool)
-
-	for _, g := range widget.RawGroups {
-		hasAny := false
-		for _, key := range g.Keys {
-			if v, ok := raw[key]; ok && v != "" {
-				hasAny = true
-				_ = v
-				break
-			}
-		}
-		if !hasAny {
-			continue
-		}
-		for _, key := range g.Keys {
-			v, ok := raw[key]
-			if !ok || v == "" {
-				continue
-			}
-			rendered[key] = true
-			fv := smartFormatValue(v)
-			if len(fv) > maxValW {
-				fv = fv[:maxValW-3] + "..."
-			}
-			sb.WriteString(fmt.Sprintf("  %s  %s\n",
-				labelStyle.Width(labelW).Render(prettifyKey(key)),
-				valueStyle.Render(fv),
-			))
-		}
-	}
-
-	keys := core.SortedStringKeys(raw)
-
-	for _, k := range keys {
-		if rendered[k] || strings.HasSuffix(k, "_error") {
-			continue
-		}
-		v := smartFormatValue(raw[k])
-		if len(v) > maxValW {
-			v = v[:maxValW-3] + "..."
-		}
-		sb.WriteString(fmt.Sprintf("  %s  %s\n",
-			labelStyle.Width(labelW).Render(prettifyKey(k)),
-			dimStyle.Render(v),
-		))
-	}
-}
-
-func smartFormatValue(v string) string {
-	trimmed := strings.TrimSpace(v)
-
-	if n, err := strconv.ParseInt(trimmed, 10, 64); err == nil && n > 1e12 && n < 2e13 {
-		t := time.Unix(n/1000, 0)
-		return t.Format("Jan 02, 2006 15:04")
-	}
-
-	if n, err := strconv.ParseInt(trimmed, 10, 64); err == nil && n > 1e9 && n < 2e10 {
-		t := time.Unix(n, 0)
-		return t.Format("Jan 02, 2006 15:04")
-	}
-
-	return v
-}
-
-func renderDetailSectionHeader(sb *strings.Builder, title string, w int) {
-	icon := sectionIcon(title)
-	sc := sectionColor(title)
-
-	iconStyled := lipgloss.NewStyle().Foreground(sc).Render(icon)
-	titleStyled := lipgloss.NewStyle().Bold(true).Foreground(sc).Render(" " + title + " ")
-	left := "  " + iconStyled + titleStyled
-
-	lineLen := w - lipgloss.Width(left) - 2
-	if lineLen < 4 {
-		lineLen = 4
-	}
-	line := lipgloss.NewStyle().Foreground(sc).Render(strings.Repeat("─", lineLen))
-	sb.WriteString(left + line + "\n")
-}
-
-func sectionIcon(title string) string {
-	switch title {
-	case "Usage":
-		return "⚡"
-	case "Spending":
-		return "💰"
-	case "Tokens":
-		return "📊"
-	case "Activity":
-		return "📈"
-	case "Timers":
-		return "⏰"
-	case "Models":
-		return "🤖"
-	case "Languages":
-		return "🗂"
-	case "Trends":
-		return "📈"
-	case "MCP Usage":
-		return "🔌"
-	case "Attributes":
-		return "📋"
-	case "Diagnostics":
-		return "⚠"
-	case "Raw Data":
-		return "🔧"
-	default:
-		return "›"
-	}
-}
-
-func sectionColor(title string) lipgloss.Color {
-	switch title {
-	case "Usage":
-		return colorYellow
-	case "Spending":
-		return colorTeal
-	case "Tokens":
-		return colorSapphire
-	case "Activity":
-		return colorGreen
-	case "Timers":
-		return colorMaroon
-	case "Models":
-		return colorLavender
-	case "Languages":
-		return colorPeach
-	case "Trends":
-		return colorSapphire
-	case "MCP Usage":
-		return colorSky
-	case "Attributes":
-		return colorBlue
-	case "Diagnostics":
-		return colorYellow
-	case "Raw Data":
-		return colorDim
-	default:
-		return colorBlue
-	}
-}
-
-func formatUsageDetail(m core.Metric) string {
-	var parts []string
-
-	if m.Remaining != nil {
-		parts = append(parts, fmt.Sprintf("%.0f%% remaining", *m.Remaining))
-	} else if m.Used != nil && m.Limit != nil {
-		rem := *m.Limit - *m.Used
-		parts = append(parts, fmt.Sprintf("%.0f%% remaining", rem))
-	}
-
-	if m.Window != "" && m.Window != "all_time" && m.Window != "current_period" {
-		parts = append(parts, "["+m.Window+"]")
-	}
-
-	return strings.Join(parts, " ")
-}
-
-func formatMetricDetail(m core.Metric) string {
-	var parts []string
-	switch {
-	case m.Used != nil && m.Limit != nil:
-		parts = append(parts, fmt.Sprintf("%s / %s %s",
-			formatNumber(*m.Used), formatNumber(*m.Limit), m.Unit))
-	case m.Remaining != nil && m.Limit != nil:
-		parts = append(parts, fmt.Sprintf("%s / %s %s remaining",
-			formatNumber(*m.Remaining), formatNumber(*m.Limit), m.Unit))
-	case m.Used != nil:
-		parts = append(parts, fmt.Sprintf("%s %s", formatNumber(*m.Used), m.Unit))
-	case m.Remaining != nil:
-		parts = append(parts, fmt.Sprintf("%s %s remaining", formatNumber(*m.Remaining), m.Unit))
-	}
-
-	if m.Window != "" && m.Window != "all_time" && m.Window != "current_period" {
-		parts = append(parts, "["+m.Window+"]")
-	}
-
-	return strings.Join(parts, " ")
-}
-
-func formatNumber(n float64) string {
-	if n == 0 {
-		return "0"
-	}
-	abs := math.Abs(n)
-	switch {
-	case abs >= 1_000_000:
-		return fmt.Sprintf("%.1fM", n/1_000_000)
-	case abs >= 10_000:
-		return fmt.Sprintf("%.1fK", n/1_000)
-	case abs >= 1_000:
-		return fmt.Sprintf("%.0f", n)
-	case abs == math.Floor(abs):
-		return fmt.Sprintf("%.0f", n)
-	default:
-		return fmt.Sprintf("%.2f", n)
-	}
-}
-
-func formatTokens(n float64) string {
-	if n == 0 {
-		return "-"
-	}
-	return formatNumber(n)
-}
-
-func formatUSD(n float64) string {
-	if n == 0 {
-		return "-"
-	}
-	if n >= 1000 {
-		return fmt.Sprintf("$%.0f", n)
-	}
-	return fmt.Sprintf("$%.2f", n)
-}
-
-func formatDuration(d time.Duration) string {
-	if d < 0 {
-		d = 0
-	}
-	switch {
-	case d < time.Minute:
-		return fmt.Sprintf("%ds", int(d.Seconds()))
-	case d < time.Hour:
-		return fmt.Sprintf("%dm%ds", int(d.Minutes()), int(d.Seconds())%60)
-	case d < 24*time.Hour:
-		return fmt.Sprintf("%dh%dm", int(d.Hours()), int(d.Minutes())%60)
-	default:
-		return fmt.Sprintf("%dd%dh", int(d.Hours())/24, int(d.Hours())%24)
-	}
-}
-
-func prettifyKey(key string) string {
-	return core.PrettifyMetricKey(key)
-}
-
-func prettifyModelName(name string) string {
-	result := strings.ReplaceAll(name, "_", "-")
-
-	switch strings.ToLower(result) {
-	case "unattributed":
-		return "unmapped spend (missing historical mapping)"
-	case "default":
-		return "default (auto)"
-	case "composer-1":
-		return "composer-1 (agent)"
-	case "github-bugbot":
-		return "github-bugbot (auto)"
-	}
-	return result
 }
