+++ internal/tui/model.go
@@ -1,12 +1,10 @@
 package tui
 
 import (
-	"fmt"
 	"strings"
 	"time"
 
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
 	"github.com/janekbaraniewski/openusage/internal/config"
 	"github.com/janekbaraniewski/openusage/internal/core"
 	"github.com/janekbaraniewski/openusage/internal/integrations"
@@ -339,535 +337,6 @@ func (m Model) shouldUsePanelScroll() bool {
 		return false
 	}
 	return m.tileCols() == 1
-}
-
-func (m Model) View() string {
-	if m.width < 30 || m.height < 8 {
-		return lipgloss.NewStyle().
-			Foreground(colorDim).
-			Render("\n  Terminal too small. Resize to at least 30×8.")
-	}
-	if !m.hasData {
-		return m.renderSplash(m.width, m.height)
-	}
-	if m.showHelp {
-		return m.renderHelpOverlay(m.width, m.height)
-	}
-	view := m.renderDashboard()
-	if m.settings.show {
-		return m.renderSettingsModalOverlay()
-	}
-	return view
-}
-
-func (m Model) renderDashboardContent(w, contentH int) string {
-	if m.mode == modeDetail {
-		return m.renderDetailPanel(w, contentH)
-	}
-	switch m.activeDashboardView() {
-	case dashboardViewTabs:
-		return m.renderTilesTabs(w, contentH)
-	case dashboardViewSplit:
-		return m.renderSplitPanes(w, contentH)
-	case dashboardViewCompare:
-		return m.renderComparePanes(w, contentH)
-	case dashboardViewStacked:
-		return m.renderTilesSingleColumn(w, contentH)
-	default:
-		return m.renderTiles(w, contentH)
-	}
-}
-
-func (m Model) renderHeader(w int) string {
-	bolt := PulseChar(
-		lipgloss.NewStyle().Foreground(colorAccent).Bold(true).Render("⚡"),
-		lipgloss.NewStyle().Foreground(colorDim).Bold(true).Render("⚡"),
-		m.animFrame,
-	)
-	brandText := RenderGradientText("OpenUsage", m.animFrame)
-
-	tabs := m.renderScreenTabs()
-
-	spinnerStr := ""
-	if m.refreshing {
-		frame := m.animFrame % len(SpinnerFrames)
-		spinnerStr = " " + lipgloss.NewStyle().Foreground(colorAccent).Render(SpinnerFrames[frame])
-	}
-
-	ids := m.filteredIDs()
-	unmappedProviders := m.telemetryUnmappedProviders()
-
-	okCount, warnCount, errCount := 0, 0, 0
-	for _, id := range ids {
-		snap := m.snapshots[id]
-		switch snap.Status {
-		case core.StatusOK:
-			okCount++
-		case core.StatusNearLimit:
-			warnCount++
-		case core.StatusLimited, core.StatusError:
-			errCount++
-		}
-	}
-
-	var info string
-
-	if m.settings.show {
-		info = m.settingsModalInfo()
-	} else {
-		switch m.screen {
-		case screenAnalytics:
-			info = dimStyle.Render("spend analysis")
-			if m.analyticsFilter.text != "" {
-				info += " (filtered)"
-			}
-		default:
-			info = fmt.Sprintf("⊞ %d providers", len(ids))
-			if m.filter.text != "" {
-				info += " (filtered)"
-			}
-			info += " · " + m.dashboardViewStatusLabel()
-		}
-	}
-	if !m.settings.show {
-		twLabel := m.timeWindow.Label()
-		info += " · " + twLabel
-	}
-	if !m.settings.show && len(unmappedProviders) > 0 {
-		info += " · detected additional providers, check settings"
-	}
-
-	statusInfo := ""
-	if okCount > 0 {
-		dot := PulseChar("●", "◉", m.animFrame)
-		statusInfo += lipgloss.NewStyle().Foreground(colorGreen).Render(fmt.Sprintf(" %d%s", okCount, dot))
-	}
-	if warnCount > 0 {
-		dot := PulseChar("◐", "◑", m.animFrame)
-		statusInfo += lipgloss.NewStyle().Foreground(colorYellow).Render(fmt.Sprintf(" %d%s", warnCount, dot))
-	}
-	if errCount > 0 {
-		dot := PulseChar("✗", "✕", m.animFrame)
-		statusInfo += lipgloss.NewStyle().Foreground(colorRed).Render(fmt.Sprintf(" %d%s", errCount, dot))
-	}
-	if len(unmappedProviders) > 0 {
-		statusInfo += lipgloss.NewStyle().
-			Foreground(colorPeach).
-			Render(fmt.Sprintf(" ⚠ %d unmapped", len(unmappedProviders)))
-	}
-
-	infoRendered := lipgloss.NewStyle().Foreground(colorSubtext).Render(info)
-
-	left := bolt + " " + brandText + " " + tabs + statusInfo + spinnerStr
-	gap := w - lipgloss.Width(left) - lipgloss.Width(infoRendered)
-	if gap < 1 {
-		gap = 1
-	}
-
-	line := left + strings.Repeat(" ", gap) + infoRendered
-
-	sep := m.renderGradientSeparator(w)
-
-	return line + "\n" + sep
-}
-
-func (m Model) renderGradientSeparator(w int) string {
-	if w <= 0 {
-		return ""
-	}
-	sepStyle := lipgloss.NewStyle().Foreground(colorSurface1)
-	return sepStyle.Render(strings.Repeat("━", w))
-}
-
-func (m Model) renderScreenTabs() string {
-	screens := m.availableScreens()
-	if len(screens) <= 1 {
-		return ""
-	}
-	var parts []string
-	for i, screen := range screens {
-		label := screenLabelByTab[screen]
-		tabStr := fmt.Sprintf("%d:%s", i+1, label)
-		if screen == m.screen {
-			parts = append(parts, screenTabActiveStyle.Render(tabStr))
-		} else {
-			parts = append(parts, screenTabInactiveStyle.Render(tabStr))
-		}
-	}
-	return strings.Join(parts, "")
-}
-
-func (m Model) renderFooter(w int) string {
-	sep := lipgloss.NewStyle().Foreground(colorSurface1).Render(strings.Repeat("━", w))
-	statusLine := m.renderFooterStatusLine(w)
-	return sep + "\n" + statusLine
-}
-
-func (m Model) renderFooterStatusLine(w int) string {
-	searchStyle := lipgloss.NewStyle().Foreground(colorSapphire)
-
-	switch {
-	case m.settings.show:
-		if m.settings.status != "" {
-			return " " + dimStyle.Render(m.settings.status)
-		}
-		return " " + helpStyle.Render("? help")
-	case m.screen == screenAnalytics:
-		if m.analyticsFilter.active {
-			cursor := PulseChar("█", "▌", m.animFrame)
-			return " " + dimStyle.Render("search: ") + searchStyle.Render(m.analyticsFilter.text+cursor)
-		}
-		if m.analyticsFilter.text != "" {
-			return " " + dimStyle.Render("filter: ") + searchStyle.Render(m.analyticsFilter.text)
-		}
-	default:
-		if m.filter.active {
-			cursor := PulseChar("█", "▌", m.animFrame)
-			return " " + dimStyle.Render("search: ") + searchStyle.Render(m.filter.text+cursor)
-		}
-		if m.filter.text != "" {
-			return " " + dimStyle.Render("filter: ") + searchStyle.Render(m.filter.text)
-		}
-		if m.activeDashboardView() == dashboardViewTabs && m.mode == modeList {
-			return " " + dimStyle.Render("tabs view · \u2190/\u2192 switch tab · PgUp/PgDn scroll widget · Enter detail")
-		}
-		if m.activeDashboardView() == dashboardViewSplit && m.mode == modeList {
-			return " " + dimStyle.Render("split view · \u2191/\u2193 select provider · PgUp/PgDn scroll pane · Enter detail")
-		}
-		if m.activeDashboardView() == dashboardViewCompare && m.mode == modeList {
-			return " " + dimStyle.Render("compare view · \u2190/\u2192 switch provider · PgUp/PgDn scroll active pane")
-		}
-		if m.mode == modeList && m.shouldUseWidgetScroll() && m.tileOffset > 0 {
-			return " " + dimStyle.Render("widget scroll active · PgUp/PgDn · Ctrl+U/Ctrl+D")
-		}
-		if m.mode == modeList && m.shouldUsePanelScroll() && m.tileOffset > 0 {
-			return " " + dimStyle.Render("panel scroll active · PgUp/PgDn · Home/End")
-		}
-	}
-
-	if m.hasAppUpdateNotice() {
-		msg := "Update available: " + m.daemon.appUpdateCurrent + " -> " + m.daemon.appUpdateLatest
-		if action := m.appUpdateAction(); action != "" {
-			msg += " · " + action
-		}
-		if w > 2 {
-			msg = truncateToWidth(msg, w-2)
-		}
-		return " " + lipgloss.NewStyle().Foreground(colorYellow).Render(msg)
-	}
-
-	return " " + helpStyle.Render("? help")
-}
-
-func (m Model) hasAppUpdateNotice() bool {
-	return strings.TrimSpace(m.daemon.appUpdateCurrent) != "" && strings.TrimSpace(m.daemon.appUpdateLatest) != ""
-}
-
-func (m Model) appUpdateHeadline() string {
-	if !m.hasAppUpdateNotice() {
-		return ""
-	}
-	return "OpenUsage update available: " + m.daemon.appUpdateCurrent + " -> " + m.daemon.appUpdateLatest
-}
-
-func (m Model) appUpdateAction() string {
-	hint := strings.TrimSpace(m.daemon.appUpdateHint)
-	if hint == "" {
-		return ""
-	}
-	return "Run: " + hint
-}
-
-func (m Model) renderList(w, h int) string {
-	ids := m.filteredIDs()
-	if len(ids) == 0 {
-		empty := []string{
-			"",
-			dimStyle.Render("  Loading providers…"),
-			"",
-			lipgloss.NewStyle().Foreground(colorSubtext).Render("  Fetching usage and spend data."),
-		}
-		return padToSize(strings.Join(empty, "\n"), w, h)
-	}
-
-	itemHeight := 3 // each item is 3 lines (name + summary + separator)
-	visibleItems := h / itemHeight
-	if visibleItems < 1 {
-		visibleItems = 1
-	}
-
-	scrollStart := 0
-	if m.cursor >= visibleItems {
-		scrollStart = m.cursor - visibleItems + 1
-	}
-	scrollEnd := scrollStart + visibleItems
-	if scrollEnd > len(ids) {
-		scrollEnd = len(ids)
-		scrollStart = scrollEnd - visibleItems
-		if scrollStart < 0 {
-			scrollStart = 0
-		}
-	}
-
-	var lines []string
-	for i := scrollStart; i < scrollEnd; i++ {
-		id := ids[i]
-		snap := m.snapshots[id]
-		selected := i == m.cursor
-		item := m.renderListItem(snap, selected, w)
-		lines = append(lines, item)
-	}
-
-	if scrollStart > 0 {
-		arrow := lipgloss.NewStyle().Foreground(colorDim).Render("  ▲ " + fmt.Sprintf("%d more", scrollStart))
-		lines = append([]string{arrow}, lines...)
-	}
-	if scrollEnd < len(ids) {
-		arrow := lipgloss.NewStyle().Foreground(colorDim).Render("  ▼ " + fmt.Sprintf("%d more", len(ids)-scrollEnd))
-		lines = append(lines, arrow)
-	}
-
-	content := strings.Join(lines, "\n")
-	out := padToSize(content, w, h)
-	if len(ids) > visibleItems && h > 0 {
-		rendered := strings.Split(out, "\n")
-		if len(rendered) > 0 {
-			rendered[len(rendered)-1] = renderVerticalScrollBarLine(w, scrollStart, visibleItems, len(ids))
-			out = strings.Join(rendered, "\n")
-		}
-	}
-	return out
-}
-
-func (m Model) renderSplitPanes(w, h int) string {
-	if w < 70 {
-		return m.renderTilesTabs(w, h)
-	}
-
-	leftW := w / 3
-	if leftW < minLeftWidth {
-		leftW = minLeftWidth
-	}
-	if leftW > maxLeftWidth {
-		leftW = maxLeftWidth
-	}
-	if leftW > w-34 {
-		leftW = w - 34
-	}
-	if leftW < minLeftWidth || w-leftW-1 < 30 {
-		return m.renderTilesTabs(w, h)
-	}
-
-	left := m.renderList(leftW, h)
-	rightW := w - leftW - 1
-	right := m.renderWidgetPanelByIndex(m.cursor, rightW, h, m.tileOffset, true)
-	sep := renderVerticalSep(h)
-
-	return lipgloss.JoinHorizontal(lipgloss.Top, left, sep, right)
-}
-
-func (m Model) renderComparePanes(w, h int) string {
-	ids := m.filteredIDs()
-	if len(ids) == 0 {
-		return m.renderTiles(w, h)
-	}
-	if len(ids) == 1 || w < 72 {
-		return m.renderWidgetPanelByIndex(m.cursor, w, h, m.tileOffset, true)
-	}
-
-	gapW := tileGapH
-	colW := (w - gapW) / 2
-	if colW < 30 {
-		return m.renderWidgetPanelByIndex(m.cursor, w, h, m.tileOffset, true)
-	}
-
-	primary := clamp(m.cursor, 0, len(ids)-1)
-	secondary := primary + 1
-	if secondary >= len(ids) {
-		secondary = primary - 1
-	}
-	if secondary < 0 {
-		secondary = primary
-	}
-
-	left := m.renderWidgetPanelByIndex(primary, colW, h, m.tileOffset, true)
-	right := m.renderWidgetPanelByIndex(secondary, colW, h, 0, false)
-
-	row := lipgloss.JoinHorizontal(lipgloss.Top, left, strings.Repeat(" ", gapW), right)
-	return padToSize(row, w, h)
-}
-
-func (m Model) renderWidgetPanelByIndex(index, w, h, bodyOffset int, selected bool) string {
-	ids := m.filteredIDs()
-	if len(ids) == 0 || index < 0 || index >= len(ids) {
-		return padToSize("", w, h)
-	}
-
-	id := ids[index]
-	snap := m.snapshots[id]
-	modelMixExpanded := index == m.cursor && m.expandedModelMixTiles[id]
-
-	tileW := w - 2 - tileBorderH
-	if tileW < tileMinWidth {
-		tileW = tileMinWidth
-	}
-	contentH := h - tileBorderV
-	if contentH < tileMinHeight {
-		contentH = tileMinHeight
-	}
-
-	rendered := m.renderTile(snap, selected, modelMixExpanded, tileW, contentH, bodyOffset)
-	return normalizeAnsiBlock(rendered, w, h)
-}
-
-func (m Model) renderListItem(snap core.UsageSnapshot, selected bool, w int) string {
-	di := computeDisplayInfo(snap, dashboardWidget(snap.ProviderID))
-
-	icon := StatusIcon(snap.Status)
-	iconColor := StatusColor(snap.Status)
-	iconStr := lipgloss.NewStyle().Foreground(iconColor).Render(icon)
-
-	nameStyle := lipgloss.NewStyle().Foreground(colorText)
-	if selected {
-		nameStyle = nameStyle.Bold(true).Foreground(colorLavender)
-	}
-
-	badge := StatusBadge(snap.Status)
-	var tagRendered string
-	if di.tagEmoji != "" && di.tagLabel != "" {
-		tc := tagColor(di.tagLabel)
-		tagRendered = lipgloss.NewStyle().Foreground(tc).Render(di.tagEmoji+" "+di.tagLabel) + " "
-	}
-	rightPart := tagRendered + badge
-	rightW := lipgloss.Width(rightPart)
-
-	name := snap.AccountID
-	maxName := w - rightW - 6 // icon + spaces + gap
-	if maxName < 5 {
-		maxName = 5
-	}
-	if len(name) > maxName {
-		name = name[:maxName-1] + "…"
-	}
-
-	namePart := fmt.Sprintf(" %s %s", iconStr, nameStyle.Render(name))
-	nameW := lipgloss.Width(namePart)
-	gapLen := w - nameW - rightW - 1
-	if gapLen < 1 {
-		gapLen = 1
-	}
-	line1 := namePart + strings.Repeat(" ", gapLen) + rightPart
-
-	summary := di.summary
-	summaryStyle := lipgloss.NewStyle().Foreground(colorText).Bold(true)
-
-	miniGauge := ""
-	if di.gaugePercent >= 0 && w > 25 {
-		gaugeW := 8
-		if w < 35 {
-			gaugeW = 5
-		}
-		miniGauge = " " + RenderMiniGauge(di.gaugePercent, gaugeW)
-	}
-
-	summaryMaxW := w - 5 - lipgloss.Width(miniGauge)
-	if summaryMaxW < 5 {
-		summaryMaxW = 5
-	}
-	if len(summary) > summaryMaxW {
-		summary = summary[:summaryMaxW-1] + "…"
-	}
-
-	line2 := "   " + summaryStyle.Render(summary) + miniGauge
-
-	line3 := "  " + lipgloss.NewStyle().Foreground(colorSurface1).Render(strings.Repeat("─", w-4))
-
-	result := line1 + "\n" + line2 + "\n" + line3
-
-	if selected {
-		indicator := lipgloss.NewStyle().Foreground(colorAccent).Bold(true).Render("┃")
-		rlines := strings.Split(result, "\n")
-		for i, l := range rlines {
-			if len(l) > 0 {
-				rlines[i] = indicator + l[1:]
-			}
-		}
-		result = strings.Join(rlines, "\n")
-	}
-
-	return result
-}
-
-func (m Model) renderDetailPanel(w, h int) string {
-	ids := m.filteredIDs()
-	if len(ids) == 0 || m.cursor >= len(ids) {
-		return padToSize("", w, h)
-	}
-
-	snap := m.snapshots[ids[m.cursor]]
-
-	tabs := DetailTabs(snap)
-	activeTab := m.detailTab
-	if activeTab >= len(tabs) {
-		activeTab = len(tabs) - 1
-	}
-	if activeTab < 0 {
-		activeTab = 0
-	}
-
-	content := m.cachedDetailContent(ids[m.cursor], snap, w-2, activeTab)
-
-	lines := strings.Split(content, "\n")
-	totalLines := len(lines)
-
-	offset := m.detailOffset
-	if offset > totalLines-h {
-		offset = totalLines - h
-	}
-	if offset < 0 {
-		offset = 0
-	}
-
-	end := offset + h
-	if end > totalLines {
-		end = totalLines
-	}
-
-	visible := lines[offset:end]
-
-	for len(visible) < h {
-		visible = append(visible, "")
-	}
-
-	result := strings.Join(visible, "\n")
-
-	if m.mode == modeDetail {
-		rlines := strings.Split(result, "\n")
-		if offset > 0 && len(rlines) > 0 {
-			arrow := lipgloss.NewStyle().Foreground(colorAccent).Render("  ▲ scroll up")
-			rlines[0] = arrow
-		}
-		if len(rlines) > 1 {
-			if bar := renderVerticalScrollBarLine(w-2, offset, h, totalLines); bar != "" {
-				rlines[len(rlines)-1] = bar
-			} else if end < totalLines {
-				arrow := lipgloss.NewStyle().Foreground(colorAccent).Render("  ▼ more below")
-				rlines[len(rlines)-1] = arrow
-			}
-		}
-		result = strings.Join(rlines, "\n")
-	}
-
-	return lipgloss.NewStyle().Width(w).Padding(0, 1).Render(result)
-}
-
-func renderVerticalSep(h int) string {
-	style := lipgloss.NewStyle().Foreground(colorSurface1)
-	lines := make([]string, h)
-	for i := range lines {
-		lines[i] = style.Render("┃")
-	}
-	return strings.Join(lines, "\n")
 }
 
 func (m *Model) applyDashboardConfig(dashboardCfg config.DashboardConfig, accounts []core.AccountConfig) {
