+++ internal/tui/settings_modal.go
@@ -2,11 +2,6 @@ package tui
 
 import (
 	"fmt"
-	"os"
-	"strings"
-
-	"github.com/charmbracelet/lipgloss"
-	"github.com/janekbaraniewski/openusage/internal/core"
 )
 
 type settingsModalTab int
@@ -90,627 +85,4 @@ func (m Model) settingsModalInfo() string {
 		info += " · " + m.settings.status
 	}
 	return info
-}
-
-func (m Model) renderSettingsProvidersBody(w, h int) string {
-	ids := m.settingsIDs()
-
-	enabledCount := 0
-	for _, id := range ids {
-		if m.isProviderEnabled(id) {
-			enabledCount++
-		}
-	}
-
-	lines := settingsBodyHeaderLines(
-		"Provider Visibility & Order",
-		fmt.Sprintf("%d/%d enabled · Shift+J/K reorder · Enter toggle", enabledCount, len(ids)),
-	)
-	accountW := 26
-	providerW := w - accountW - 16
-	if providerW < 10 {
-		providerW = 10
-		accountW = w - providerW - 16
-	}
-	if accountW < 12 {
-		accountW = 12
-	}
-	lines = append(lines, dimStyle.Render(fmt.Sprintf("    %-3s %-3s %-*s %-*s", "#", "ON", accountW, "ACCOUNT", providerW, "PROVIDER")))
-	lines = append(lines, settingsBodyRule(w))
-	if len(ids) == 0 {
-		lines = append(lines, dimStyle.Render("No providers available."))
-		return padToSize(strings.Join(lines, "\n"), w, h)
-	}
-
-	cursor := clamp(m.settings.cursor, 0, len(ids)-1)
-	listHeight := h - len(lines)
-	if listHeight < 1 {
-		listHeight = 1
-	}
-	start, end := listWindow(len(ids), cursor, listHeight)
-
-	for i := start; i < end; i++ {
-		id := ids[i]
-		providerID := m.accountProviders[id]
-		if snap, ok := m.snapshots[id]; ok && snap.ProviderID != "" {
-			providerID = snap.ProviderID
-		}
-		if providerID == "" {
-			providerID = "unknown"
-		}
-
-		onText := "OFF"
-		onStyle := lipgloss.NewStyle().Foreground(colorRed)
-		if m.isProviderEnabled(id) {
-			onText = "ON "
-			onStyle = lipgloss.NewStyle().Foreground(colorGreen)
-		}
-
-		prefix := "  "
-		if i == cursor {
-			prefix = lipgloss.NewStyle().Foreground(colorAccent).Bold(true).Render("➤ ")
-		}
-		account := truncateToWidth(id, accountW)
-		provider := truncateToWidth(providerID, providerW)
-		line := fmt.Sprintf("%s%-3d %s %-*s %-*s", prefix, i+1, onStyle.Render(onText), accountW, account, providerW, provider)
-		lines = append(lines, line)
-	}
-
-	return padToSize(strings.Join(lines, "\n"), w, h)
-}
-
-func (m Model) renderSettingsWidgetSectionsBody(w, h int) string {
-	return m.renderSettingsWidgetSectionsList(w, h)
-}
-
-func (m Model) renderSettingsWidgetSectionsList(w, h int) string {
-	entries := m.widgetSectionEntries()
-
-	visibleCount := 0
-	for _, entry := range entries {
-		if entry.Enabled {
-			visibleCount++
-		}
-	}
-
-	lines := settingsBodyHeaderLines(
-		"Global Widget Sections",
-		fmt.Sprintf("%d/%d sections visible · applies to all providers", visibleCount, len(entries)),
-	)
-	hideBox := "☐"
-	hideBoxStyle := lipgloss.NewStyle().Foreground(colorRed)
-	if m.hideSectionsWithNoData {
-		hideBox = "☑"
-		hideBoxStyle = lipgloss.NewStyle().Foreground(colorGreen)
-	}
-	lines = append(lines, fmt.Sprintf("Hide sections with no data: %s  %s", hideBoxStyle.Render(hideBox), dimStyle.Render("press h to toggle")))
-	lines = append(lines, "")
-	nameW := w - 24
-	if nameW < 12 {
-		nameW = 12
-	}
-	lines = append(lines, dimStyle.Render(fmt.Sprintf("    %-3s %-3s %-*s %s", "#", "ON", nameW, "SECTION", "ID")))
-	lines = append(lines, settingsBodyRule(w))
-	if len(entries) == 0 {
-		lines = append(lines, dimStyle.Render("No dashboard sections available."))
-		return padToSize(strings.Join(lines, "\n"), w, h)
-	}
-
-	cursor := clamp(m.settings.sectionRowCursor, 0, len(entries)-1)
-	listHeight := h - len(lines)
-	if listHeight < 1 {
-		listHeight = 1
-	}
-	start, end := listWindow(len(entries), cursor, listHeight)
-
-	for i := start; i < end; i++ {
-		entry := entries[i]
-		prefix := "  "
-		if i == cursor {
-			prefix = lipgloss.NewStyle().Foreground(colorAccent).Bold(true).Render("➤ ")
-		}
-
-		onText := "OFF"
-		onStyle := lipgloss.NewStyle().Foreground(colorRed)
-		if entry.Enabled {
-			onText = "ON "
-			onStyle = lipgloss.NewStyle().Foreground(colorGreen)
-		}
-
-		name := settingsSectionLabel(entry.ID)
-		name = truncateToWidth(name, nameW)
-		line := fmt.Sprintf("%s%-3d %s %-*s %s", prefix, i+1, onStyle.Render(onText), nameW, name, dimStyle.Render(string(entry.ID)))
-		lines = append(lines, line)
-	}
-
-	return padToSize(strings.Join(lines, "\n"), w, h)
-}
-
-func (m Model) renderSettingsWidgetSectionsPreview(w, h int) string {
-	if w < 24 || h < 5 {
-		return padToSize(dimStyle.Render("Live preview unavailable at this size."), w, h)
-	}
-
-	title := lipgloss.NewStyle().Foreground(colorTeal).Bold(true).Render("Live Preview")
-	hint := dimStyle.Render("Claude Code preset · synthetic data · PgUp/PgDn scroll")
-	lines := []string{title, hint, ""}
-
-	tileW := w
-	if tileW > 2 {
-		tileW -= 2
-	}
-	if tileW < tileMinWidth {
-		tileW = tileMinWidth
-	}
-
-	// Render full tile content to avoid nested-scroll artifacts inside the preview panel.
-	previewTile := m.renderTile(settingsWidgetSectionsPreviewSnapshot(), false, false, tileW, 0, 0)
-	all := append(lines, strings.Split(previewTile, "\n")...)
-	maxOffset := len(all) - h
-	if maxOffset < 0 {
-		maxOffset = 0
-	}
-	offset := clamp(m.settings.previewOffset, 0, maxOffset)
-	visible := all
-	if len(visible) > h {
-		visible = visible[offset:]
-		if len(visible) > h {
-			visible = visible[:h]
-		}
-	}
-	if len(visible) > 0 && offset > 0 {
-		visible[0] = dimStyle.Render("  ▲ preview above")
-	}
-	if len(visible) > 0 && offset+h < len(all) {
-		visible[len(visible)-1] = dimStyle.Render("  ▼ preview below")
-	}
-	return padToSize(strings.Join(visible, "\n"), w, h)
-}
-
-func (m Model) renderSettingsWidgetPreviewPanel(contentW, contentH int) string {
-	innerW := contentW - 4
-	if innerW < 24 {
-		innerW = contentW
-	}
-	bodyH := contentH - 1
-	if bodyH < 4 {
-		bodyH = 4
-	}
-	title := lipgloss.NewStyle().Bold(true).Foreground(colorRosewater).Render("Widget Preview")
-	body := m.renderSettingsWidgetSectionsPreview(innerW, bodyH)
-	lines := []string{
-		title,
-		lipgloss.NewStyle().Foreground(colorSurface1).Render(strings.Repeat("─", innerW)),
-		body,
-	}
-	return lipgloss.NewStyle().
-		Border(lipgloss.RoundedBorder()).
-		BorderForeground(colorAccent).
-		Background(colorBase).
-		Padding(1, 2, 0, 2).
-		Width(contentW).
-		Render(strings.Join(lines, "\n"))
-}
-
-func (m Model) settingsWidgetPreviewBodyHeight(contentW, contentH int, sideBySide bool) int {
-	minBodyH := settingsWidgetPreviewMinBodyH
-	maxBodyH := contentH
-	if sideBySide {
-		// Keep breathing room around the combined modal while allowing growth.
-		maxBodyH = m.height - 12
-	} else {
-		// Stacked layout should stay balanced and avoid dominating the viewport.
-		maxBodyH = (m.height - 12) / 2
-	}
-	if maxBodyH < minBodyH {
-		maxBodyH = minBodyH
-	}
-
-	innerW := contentW - 4
-	if innerW < 24 {
-		innerW = 24
-	}
-	targetBodyH := m.settingsWidgetPreviewContentLineCount(innerW)
-	if targetBodyH < minBodyH {
-		targetBodyH = minBodyH
-	}
-	if targetBodyH > maxBodyH {
-		targetBodyH = maxBodyH
-	}
-
-	// renderSettingsWidgetPreviewPanel reserves one line for panel internals.
-	return targetBodyH + 1
-}
-
-func (m Model) settingsWidgetPreviewContentLineCount(innerW int) int {
-	if innerW < 24 {
-		return 4
-	}
-	tileW := innerW
-	if tileW > 2 {
-		tileW -= 2
-	}
-	if tileW < tileMinWidth {
-		tileW = tileMinWidth
-	}
-	previewTile := m.renderTile(settingsWidgetSectionsPreviewSnapshot(), false, false, tileW, 0, 0)
-	// Includes preview title line, hint line, and spacing line.
-	return 3 + len(strings.Split(previewTile, "\n"))
-}
-
-func centerPanelVertically(panel string, targetHeight int) string {
-	current := lipgloss.Height(panel)
-	if current >= targetHeight {
-		return panel
-	}
-	diff := targetHeight - current
-	top := diff / 2
-	bottom := diff - top
-	return strings.Repeat("\n", top) + panel + strings.Repeat("\n", bottom)
-}
-
-func (m Model) renderSettingsThemeBody(w, h int) string {
-	themes := AvailableThemes()
-	activeThemeIdx := ActiveThemeIndex()
-	activeThemeName := "none"
-	if activeThemeIdx >= 0 && activeThemeIdx < len(themes) {
-		activeThemeName = themes[activeThemeIdx].Name
-	}
-	lines := settingsBodyHeaderLines(
-		"Theme Selection",
-		fmt.Sprintf("%d themes available · active: %s", len(themes), activeThemeName),
-	)
-	nameW := w - 16
-	if nameW < 12 {
-		nameW = 12
-	}
-	lines = append(lines, dimStyle.Render(fmt.Sprintf("    %-3s %-3s %-3s %-*s", "#", "CUR", "ACT", nameW, "THEME")))
-	lines = append(lines, settingsBodyRule(w))
-	if len(themes) == 0 {
-		lines = append(lines, dimStyle.Render("No themes available."))
-		return padToSize(strings.Join(lines, "\n"), w, h)
-	}
-
-	cursor := clamp(m.settings.themeCursor, 0, len(themes)-1)
-	listHeight := h - len(lines)
-	if listHeight < 1 {
-		listHeight = 1
-	}
-	start, end := listWindow(len(themes), cursor, listHeight)
-
-	for i := start; i < end; i++ {
-		theme := themes[i]
-		prefix := "  "
-		if i == cursor {
-			prefix = lipgloss.NewStyle().Foreground(colorAccent).Bold(true).Render("➤ ")
-		}
-
-		current := "."
-		if i == activeThemeIdx {
-			current = "*"
-		}
-		selected := "."
-		if i == cursor {
-			selected = ">"
-		}
-		name := truncateToWidth(theme.Name, nameW)
-		lines = append(lines, fmt.Sprintf("%s%-3d %-3s %-3s %-*s", prefix, i+1, selected, current, nameW, name))
-	}
-
-	return padToSize(strings.Join(lines, "\n"), w, h)
-}
-
-func (m Model) renderSettingsViewBody(w, h int) string {
-	configured := m.configuredDashboardView()
-	active := m.activeDashboardView()
-	lines := settingsBodyHeaderLines(
-		"Dashboard View Mode",
-		fmt.Sprintf("configured: %s · active: %s", configured, active),
-	)
-	lines = append(lines, dimStyle.Render("    CUR  MODE"))
-	lines = append(lines, settingsBodyRule(w))
-	if len(dashboardViewOptions) == 0 {
-		lines = append(lines, dimStyle.Render("No dashboard views available."))
-		return padToSize(strings.Join(lines, "\n"), w, h)
-	}
-
-	cursor := clamp(m.settings.viewCursor, 0, len(dashboardViewOptions)-1)
-	listHeight := h - len(lines)
-	if listHeight < 1 {
-		listHeight = 1
-	}
-	start, end := listWindow(len(dashboardViewOptions), cursor, listHeight)
-
-	for i := start; i < end; i++ {
-		option := dashboardViewOptions[i]
-
-		prefix := "  "
-		if i == cursor {
-			prefix = lipgloss.NewStyle().Foreground(colorAccent).Bold(true).Render("➤ ")
-		}
-
-		current := "  "
-		if option.ID == configured {
-			current = lipgloss.NewStyle().Foreground(colorGreen).Bold(true).Render("● ")
-		}
-
-		label := option.Label
-		if option.ID == active && option.ID != configured {
-			label += " (auto)"
-		}
-
-		lines = append(lines, fmt.Sprintf("%s%s%s", prefix, current, label))
-		lines = append(lines, "    "+dimStyle.Render(option.Description))
-	}
-
-	return padToSize(strings.Join(lines, "\n"), w, h)
-}
-
-// apiKeysTabIDs returns account IDs for the API Keys tab, including
-// unregistered API-key providers that the user can configure.
-func (m Model) apiKeysTabIDs() []string {
-	registeredProviders := make(map[string]bool)
-	var ids []string
-	for _, id := range m.providerOrder {
-		providerID := m.accountProviders[id]
-		if isAPIKeyProvider(providerID) {
-			ids = append(ids, id)
-			registeredProviders[providerID] = true
-		}
-	}
-	for _, entry := range apiKeyProviderEntries() {
-		if registeredProviders[entry.ProviderID] {
-			continue
-		}
-		ids = append(ids, entry.AccountID)
-	}
-	return ids
-}
-
-// providerForAccountID looks up the provider ID for an account, falling back
-// to the default API-key account mapping for unregistered providers.
-func providerForAccountID(accountID string, accountProviders map[string]string) string {
-	if p, ok := accountProviders[accountID]; ok && p != "" {
-		return p
-	}
-	for _, entry := range apiKeyProviderEntries() {
-		if entry.AccountID == accountID {
-			return entry.ProviderID
-		}
-	}
-	return ""
-}
-
-func maskAPIKey(key string) string {
-	if len(key) <= 12 {
-		return key
-	}
-	return key[:8] + "..." + key[len(key)-4:]
-}
-
-func (m Model) renderSettingsAPIKeysBody(w, h int) string {
-	ids := m.apiKeysTabIDs()
-
-	configuredCount := 0
-	for _, id := range ids {
-		providerID := providerForAccountID(id, m.accountProviders)
-		if !isAPIKeyProvider(providerID) {
-			continue
-		}
-		if envVar := envVarForProvider(providerID); envVar != "" && os.Getenv(envVar) != "" {
-			configuredCount++
-			continue
-		}
-		if snap, ok := m.snapshots[id]; ok && snap.Status == core.StatusOK {
-			configuredCount++
-		}
-	}
-
-	lines := settingsBodyHeaderLines(
-		"API Key Management",
-		fmt.Sprintf("%d/%d configured (env or validated)", configuredCount, len(ids)),
-	)
-	accountW := 20
-	envW := w - accountW - 18
-	if envW < 10 {
-		envW = 10
-		accountW = w - envW - 18
-	}
-	if accountW < 10 {
-		accountW = 10
-	}
-	lines = append(lines, dimStyle.Render(fmt.Sprintf("    %-3s %-5s %-*s %-*s", "#", "STAT", accountW, "ACCOUNT", envW, "ENV VAR")))
-	lines = append(lines, settingsBodyRule(w))
-	if len(ids) == 0 {
-		lines = append(lines, dimStyle.Render("No API-key providers available."))
-		return padToSize(strings.Join(lines, "\n"), w, h)
-	}
-
-	cursor := clamp(m.settings.cursor, 0, len(ids)-1)
-	listHeight := h - len(lines)
-	if listHeight < 1 {
-		listHeight = 1
-	}
-	start, end := listWindow(len(ids), cursor, listHeight)
-
-	for i := start; i < end; i++ {
-		id := ids[i]
-		providerID := providerForAccountID(id, m.accountProviders)
-		if snap, ok := m.snapshots[id]; ok && snap.ProviderID != "" {
-			providerID = snap.ProviderID
-		}
-		if providerID == "" {
-			providerID = "unknown"
-		}
-
-		prefix := "  "
-		if i == cursor {
-			prefix = lipgloss.NewStyle().Foreground(colorAccent).Bold(true).Render("➤ ")
-		}
-
-		if !isAPIKeyProvider(providerID) {
-			line := fmt.Sprintf("%s%-3d %-5s %-*s %-*s", prefix, i+1, "N/A", accountW, truncateToWidth(id, accountW), envW, "-")
-			lines = append(lines, line)
-			continue
-		}
-
-		envVar := envVarForProvider(providerID)
-
-		var statusText string
-		if snap, ok := m.snapshots[id]; ok && snap.Status == core.StatusOK {
-			statusText = "OK"
-		} else if envVar != "" && os.Getenv(envVar) != "" {
-			statusText = "ENV"
-		} else {
-			statusText = "MISS"
-		}
-
-		account := truncateToWidth(id, accountW)
-		envLabel := "-"
-		if envVar != "" {
-			envLabel = envVar
-		}
-		envLabel = truncateToWidth(envLabel, envW)
-
-		if m.settings.apiKeyEditing && i == cursor {
-			masked := maskAPIKey(m.settings.apiKeyInput)
-			inputStyle := lipgloss.NewStyle().Foreground(colorSapphire)
-			cursorChar := PulseChar("█", "▌", m.animFrame)
-			line := fmt.Sprintf("%s%-3d %-5s %-*s %-*s", prefix, i+1, statusText, accountW, account, envW, envLabel)
-			lines = append(lines, line)
-			keyLine := fmt.Sprintf("     key: %s", inputStyle.Render(masked+cursorChar))
-			if m.settings.apiKeyStatus != "" {
-				keyLine += "  " + dimStyle.Render(m.settings.apiKeyStatus)
-			}
-			lines = append(lines, keyLine)
-		} else {
-			line := fmt.Sprintf("%s%-3d %-5s %-*s %-*s", prefix, i+1, statusText, accountW, account, envW, envLabel)
-			lines = append(lines, line)
-		}
-	}
-
-	return padToSize(strings.Join(lines, "\n"), w, h)
-}
-
-func (m Model) renderSettingsTelemetryBody(w, h int) string {
-	lines := settingsBodyHeaderLines(
-		"Telemetry & Time Window",
-		"Choose aggregation window and map raw telemetry providers",
-	)
-	lines = append(lines, settingsBodyRule(w))
-	lines = append(lines, "")
-
-	// Time window selector
-	lines = append(lines, lipgloss.NewStyle().Foreground(colorTeal).Bold(true).Render("Time Window")+"  "+dimStyle.Render("press w or select below"))
-	lines = append(lines, "")
-	for i, tw := range core.ValidTimeWindows {
-		prefix := "  "
-		if i == m.settings.cursor {
-			prefix = lipgloss.NewStyle().Foreground(colorAccent).Bold(true).Render("➤ ")
-		}
-		current := "  "
-		if tw == m.timeWindow {
-			current = lipgloss.NewStyle().Foreground(colorGreen).Bold(true).Render("● ")
-		}
-		lines = append(lines, fmt.Sprintf("%s%s%s", prefix, current, tw.Label()))
-	}
-	lines = append(lines, "")
-
-	// Telemetry provider mapping section
-	unmapped := m.telemetryUnmappedProviders()
-	hints := m.telemetryProviderLinkHints()
-	configured := m.configuredProviderIDs()
-
-	if len(unmapped) == 0 {
-		lines = append(lines, lipgloss.NewStyle().Foreground(colorGreen).Render("All telemetry providers are mapped."))
-	} else {
-		lines = append(lines, lipgloss.NewStyle().Foreground(colorPeach).Bold(true).Render("Detected additional telemetry providers:"))
-		for _, providerID := range unmapped {
-			lines = append(lines, "  - "+providerID)
-		}
-		lines = append(lines, "")
-		lines = append(lines, "Map them in settings.json under telemetry.provider_links:")
-		lines = append(lines, "  <source_provider>=<configured_provider_id>")
-		if len(hints) > 0 {
-			lines = append(lines, "")
-			lines = append(lines, "Hint:")
-			lines = append(lines, "  "+hints[0])
-		}
-		if len(configured) > 0 {
-			lines = append(lines, "")
-			lines = append(lines, "Configured provider IDs:")
-			lines = append(lines, "  "+strings.Join(configured, ", "))
-		}
-	}
-
-	start, end := listWindow(len(lines), m.settings.bodyOffset, h)
-	return padToSize(strings.Join(lines[start:end], "\n"), w, h)
-}
-
-func (m Model) renderSettingsIntegrationsBody(w, h int) string {
-	statuses := m.settings.integrationStatus
-	ready := 0
-	outdated := 0
-	for _, entry := range statuses {
-		if entry.State == "ready" {
-			ready++
-		}
-		if entry.NeedsUpgrade || entry.State == "outdated" {
-			outdated++
-		}
-	}
-	lines := settingsBodyHeaderLines(
-		"Integrations",
-		fmt.Sprintf("%d total · %d ready · %d need attention", len(statuses), ready, outdated),
-	)
-	lines = append(lines, settingsBodyRule(w))
-	if len(statuses) == 0 {
-		lines = append(lines, dimStyle.Render("No integration status available yet. Press r to refresh."))
-		return padToSize(strings.Join(lines, "\n"), w, h)
-	}
-
-	cursor := clamp(m.settings.cursor, 0, len(statuses)-1)
-	listHeight := h - len(lines) - 4
-	if listHeight < 1 {
-		listHeight = 1
-	}
-	start, end := listWindow(len(statuses), cursor, listHeight)
-
-	for i := start; i < end; i++ {
-		entry := statuses[i]
-		prefix := "  "
-		if i == cursor {
-			prefix = lipgloss.NewStyle().Foreground(colorAccent).Bold(true).Render("➤ ")
-		}
-
-		stateColor := colorRed
-		switch entry.State {
-		case "ready":
-			stateColor = colorGreen
-		case "outdated":
-			stateColor = colorYellow
-		case "partial":
-			stateColor = colorPeach
-		}
-
-		versionText := entry.DesiredVersion
-		if strings.TrimSpace(entry.InstalledVersion) != "" {
-			versionText = entry.InstalledVersion
-		}
-		stateText := lipgloss.NewStyle().Foreground(stateColor).Render(strings.ToUpper(entry.State))
-		line := fmt.Sprintf("%s%s  %s  %s", prefix, entry.Name, stateText, dimStyle.Render("v"+versionText))
-		lines = append(lines, line)
-		lines = append(lines, "    "+dimStyle.Render(entry.Summary))
-	}
-
-	selected := statuses[cursor]
-	lines = append(lines, "")
-	lines = append(lines, "Selected:")
-	lines = append(lines, fmt.Sprintf("  %s · installed=%t configured=%t", selected.Name, selected.Installed, selected.Configured))
-	if selected.NeedsUpgrade {
-		lines = append(lines, "  "+lipgloss.NewStyle().Foreground(colorYellow).Render("Upgrade recommended: installed version differs from current integration version"))
-	}
-	lines = append(lines, "  Install/configure command writes plugin/hook files and updates tool configs automatically.")
-
-	return padToSize(strings.Join(lines, "\n"), w, h)
 }
